modify edge preditor, pkl
