+++ logicity/utils/dataset.py
@@ -3,6 +3,8 @@ import torch
 import pickle as pkl
 from PIL import Image, ImageDraw, ImageFont
 import os
+import cv2
+
 
 # data loader class for data buffer inherited from torch.utils.data.Dataset
 class WMDataset(torch.utils.data.Dataset):
@@ -62,12 +64,12 @@ class VisDataset(torch.utils.data.Dataset):
             vis_dataset_path, 
             batch_size=1, 
             shuffle=True,
-            bbox_vis=False,
+            gt_vis=False,
         ):
 
         self.batch_size = batch_size
         self.shuffle = shuffle
-        self.bbox_vis = bbox_vis
+        self.gt_vis = gt_vis
         with open(vis_dataset_path, "rb") as f:
             self.vis_dataset = pkl.load(f)
         self.vis_dataset_list = list(self.vis_dataset.keys())
@@ -88,7 +90,7 @@ class VisDataset(torch.utils.data.Dataset):
         img = np.array(img)
         return img
     
-    def bboxes_visualization(self, step_name, img_path, bboxes, detailed_types, output_folder):
+    def gt_visualization(self, step_name, img_path, predicates, bboxes, detailed_types, output_folder):
         im = Image.open(img_path)
         draw = ImageDraw.Draw(im)
         # Define font type and size (you might need to provide the path to a .ttf font file)
@@ -104,9 +106,35 @@ class VisDataset(torch.utils.data.Dataset):
             draw.rectangle(bbox, outline='red',width=2)
             draw.rectangle(label_bbox, fill='red')
             draw.text(text_origin, label, fill=(255, 255, 255), font=font)
+
+        na = np.array(im)
+        agent_num = len(bboxes)
+        bbox_centers = []
+        for i in range(agent_num):
+            bbox_centers.append([(bboxes[i][0]+bboxes[i][2])//2, (bboxes[i][1]+bboxes[i][3])//2])
+        for i in range(agent_num):
+            for j in range(agent_num):
+                if i == j:
+                    continue
+                p_str = ""
+                for p_name in predicates.keys():
+                    if "Is" not in p_name and predicates[p_name][i][j]:
+                        p_str += p_name + ", "
+                if len(p_str) != 0:
+                    if i > j:
+                        start_pos = (bbox_centers[i][0]-5, bbox_centers[i][1]-5)
+                        end_pos = (bbox_centers[j][0]-5, bbox_centers[j][1]-5)
+                    else:
+                        start_pos = (bbox_centers[i][0]+5, bbox_centers[i][1]+5)
+                        end_pos = (bbox_centers[j][0]+5, bbox_centers[j][1]+5)
+                    text_pos = ((start_pos[0]*2//3+end_pos[0]//3)-10, (start_pos[1]*2//3+end_pos[1]//3))
+                    cv2.arrowedLine(na, start_pos, end_pos, (0, 255, 0), 1, tipLength = 0.03)
+                    cv2.putText(na, p_str, text_pos, cv2.FONT_HERSHEY_PLAIN, 1, (0, 255, 0), 1, cv2.LINE_AA)
+
         if not os.path.exists(output_folder):
             os.makedirs(output_folder)
         output_path = "{}/{}.png".format(output_folder, step_name)
+        im = Image.fromarray(na)
         im.save(output_path)
         return im
 
@@ -116,6 +144,7 @@ class VisDataset(torch.utils.data.Dataset):
     def __getitem__(self, idx):
         step_names = self.vis_dataset_list[idx*self.batch_size:(idx+1)*self.batch_size]
         imgs = []
+        predicates = []
         bboxes = []
         types = []
         detailed_types = []
@@ -124,6 +153,7 @@ class VisDataset(torch.utils.data.Dataset):
         next_actions = []
         for step_name in step_names:
             imgs.append(self.read_img(self.vis_dataset[step_name]["Image_path"]))
+            predicates.append(self.vis_dataset[step_name]["Predicate_groundings"])
             bboxes.append(list(self.vis_dataset[step_name]["Bboxes"].values()))
             types.append(list(self.vis_dataset[step_name]["Types"].values()))
             detailed_types.append(list(self.vis_dataset[step_name]["Detailed_types"].values()))
@@ -135,9 +165,9 @@ class VisDataset(torch.utils.data.Dataset):
             directions.append(direction_tensor_list)
             next_actions.append(list(self.vis_dataset[step_name]["Next_actions"].values()))
 
-            if self.bbox_vis:
-                output_folder = "bbox_vis/{}".format(step_name.split('_')[0])
-                self.bboxes_visualization(step_name, self.vis_dataset[step_name]["Image_path"], bboxes[-1], types[-1], output_folder)
+            if self.gt_vis:
+                output_folder = "gt_vis/{}".format(step_name.split('_')[0])
+                self.gt_visualization(step_name, self.vis_dataset[step_name]["Image_path"], predicates[0], bboxes[0], types[0], output_folder)
 
         imgs = torch.Tensor(np.array(imgs))
         bboxes = torch.Tensor(np.array(bboxes))
@@ -149,6 +179,7 @@ class VisDataset(torch.utils.data.Dataset):
         out_dict = {
             "step_names": step_names,
             "imgs": imgs,
+            "predicates": predicates,
             "bboxes": bboxes,
             "types": types,
             "detailed_types": detailed_types,
